- Remove notification.js - Modify Notifier.jsx so it accepts toast type and apply defaultStyles - Replace old notify with new notify from Notifier.jsx
