+++ frontend/src/App.jsx
@@ -13,12 +13,11 @@ import { getTelegramUserWalletId } from 'services/telegram';
 import Documentation from 'pages/spotnet/documentation/Documentation';
 import Withdraw from 'pages/vault/withdraw/Withdraw';
 import { useWalletStore } from 'stores/useWalletStore';
-import { Notifier } from 'components/Notifier/Notifier';
+import { Notifier, notify } from 'components/Notifier/Notifier';
 import { useConnectWallet } from 'hooks/useConnectWallet';
 import OverviewPage from 'pages/spotnet/overview/Overview';
 import { ActionModal } from 'components/ui/ActionModal';
 import Stake from 'pages/vault/stake/Stake';
-import { notifyError } from 'utils/notification';
 
 function App() {
   const { walletId, setWalletId, removeWalletId } = useWalletStore();
@@ -54,7 +53,7 @@ function App() {
         window.Telegram.WebApp.ready();
       }).catch((error) => {
         console.error('Error getting Telegram user wallet ID:', error);
-        notifyError('Error loading wallet');
+        notify('Error loading wallet', "error");
         window.Telegram.WebApp.ready();
       });
     }
