+++ frontend/src/components/Notifier/Notifier.jsx
@@ -2,7 +2,14 @@ import React from 'react';
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
-const notify = (message, id, autoClose) => toast(message, {toastId: id || undefined, autoClose: autoClose !== undefined ? autoClose : 3000});
+const defaultStyles = {
+  success: { backgroundColor: 'green', color: 'white' },
+  error: { backgroundColor: 'red', color: 'white' },
+  warning: { backgroundColor: 'orange', color: 'white' },
+  info: { backgroundColor: 'blue', color: 'white' },
+};
+
+const notify = (message, type='info', autoClose=3000) => toast(message, { type, autoClose, style: defaultStyles[type] || defaultStyles.info });
 
 const Notifier = () => {
   return (
